requirements.txt
matplotlib
numpy
scipy
networkx
stim
sinter>=1.14
pycollada
typing-extensions
pycryptosat!=5.11.23
python-sat[cryptosat]
tqecd>=0.1.2
pyzx>=0.9.0
platformdirs
svg.py
